% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qsub.R
\name{create.qsub.object}
\alias{create.qsub.object}
\title{Create a qsub configuration object.}
\usage{
create.qsub.object(remote, qsub.src.dir, qsub.remote.dir, qsub.tmp.foldername)
}
\arguments{
\item{remote}{Remote machine specification for ssh, in format such as \code{user@server} that does not
require interactive password entry.}

\item{qsub.src.dir}{A directory on the local machine in which to store temporary files.}

\item{qsub.remote.dir}{A directory on the remote machine in which to store temporary files.}

\item{qsub.tmp.foldername}{A temporary name for this execution.}
}
\description{
Create a qsub configuration object.
}

