% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub.R
\name{create.server.config}
\alias{create.server.config}
\title{Create a new server config file}
\usage{
create.server.config(remote, src.dir, remote.dir,
  file = "~/.local/share/R2PRISM/config")
}
\arguments{
\item{remote}{Remote machine specification for ssh, in format such as \code{user@server} that does not
require interactive password entry.}

\item{src.dir}{A directory on the local machine in which to store temporary files.}

\item{remote.dir}{A directory on the remote machine in which to store temporary files.}

\item{file}{The default location to store the R2PRISM config file.}
}
\description{
Create a new server config file
}
\examples{
create.server.config("myservername", "~/tmpfolderlocal", "~/tmpfolder_remote")
}

