% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub_config.R
\name{create_qsub_config}
\alias{create_qsub_config}
\title{Create a qsub configuration object.}
\usage{
create_qsub_config(remote, local_tmp_path, remote_tmp_path, r_module = "R",
  name = "R2PRISM", num_cores = 1, max_running_tasks = NULL,
  memory = "4G", verbose = F, wait = T, remove_tmp_folder = T,
  stop_on_error = T, max_tasks = NULL, execute_before = NULL)
}
\arguments{
\item{remote}{Remote machine specification for ssh, in format such as \code{user@server}
that does not require interactive password entry.}

\item{local_tmp_path}{A directory on the local machine in which to store temporary files. Should not contain a tilde ('~').}

\item{remote_tmp_path}{A directory on the remote machine in which to store temporary files. Should not contain a tilde ('~').}

\item{r_module}{The R module to use}

\item{name}{The name of the execution}

\item{num_cores}{The number of cores to allocate per task}

\item{max_running_tasks}{öimit concurrent array job task execution}

\item{memory}{The memory to allocate per core}

\item{verbose}{Print out any ssh commands}

\item{wait}{If \code{TRUE}, will wait until the execution has finished by periodically checking the job status.}

\item{remove_tmp_folder}{If \code{TRUE}, will remove everything that was created related to this execution at the end.}

\item{stop_on_error}{If \code{TRUE}, will stop when an error occurs, else returns a NA for errored instances.}

\item{max_tasks}{The maximum number of tasks to spawn}

\item{execute_before}{Commands to execute in the shell before running R}
}
\value{
A qsub configuration object.
}
\description{
Create a qsub configuration object.
}
\examples{
\dontrun{
qsub_config <- create_qsub_config(
  remote = "myserver",
  local_tmp_path = "/home/myuser/workspace/.r2gridengine",
  remote_tmp_path = "/scratch/myuser/.r2gridengine"
)
qsub_lapply(1:10, function(x) x + 1, qsub_config = qsub_config)

set_default_qsub_config(qsub_config, permanent = T)
qsub_lapply(1:10, function(x) x + 1)
}
}
\seealso{
\code{\link{qsub_lapply}}, \code{\link{set_default_qsub_config}}
}

