% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub_config.R
\name{get_default_qsub_config}
\alias{get_default_qsub_config}
\title{Retrieve a default qsub_config.}
\usage{
get_default_qsub_config(permanent_file = "~/.local/share/R2PRISM/qsub_config.rds")
}
\arguments{
\item{permanent_file}{The file in which a permanent default config is stored.}
}
\description{
Will prefer the temporary default over the permanent default.
You should typically not require this function.
}
