% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub_config.R
\name{instantiate_qsub_config}
\alias{instantiate_qsub_config}
\title{Create an instance of the qsub_config.}
\usage{
instantiate_qsub_config(qsub_config)
}
\arguments{
\item{qsub_config}{A valid qsub_config object.

## @export # you should typically not require to call this function manually.}
}
\description{
This function generates the paths for the temporary files.
}
