% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub.R
\name{is.job.running}
\alias{is.job.running}
\title{Check whether a job is running.}
\usage{
is.job.running(qsub.config)
}
\arguments{
\item{qsub.config}{The qsub configuration of class \code{qsub_configuration}, as returned by any qsub execution}
}
\description{
Check whether a job is running.
}

