% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub.R
\name{is.qsub.config}
\alias{is.qsub.config}
\title{Outputs whether the object is a qsub configuration.}
\usage{
is.qsub.config(config)
}
\arguments{
\item{config}{}
}
\description{
Outputs whether the object is a qsub configuration.
}

