% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub.R
\name{is.server.config}
\alias{is.server.config}
\title{Outputs whether the object is a server configuration.}
\usage{
is.server.config(config)
}
\arguments{
\item{config}{}
}
\description{
Outputs whether the object is a server configuration.
}

