% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub_run.R
\name{is_job_running}
\alias{is_job_running}
\title{Check whether a job is running.}
\usage{
is_job_running(qsub_config)
}
\arguments{
\item{qsub_config}{The qsub configuration of class \code{qsub_configuration}, as returned by any qsub execution}
}
\description{
Check whether a job is running.
}
