% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub_config.R
\name{is_qsub_config}
\alias{is_qsub_config}
\title{Returns whether the passed object is a qsub_config object.}
\usage{
is_qsub_config(object)
}
\arguments{
\item{object}{The object to be tested}
}
\description{
Returns whether the passed object is a qsub_config object.
}

