% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{is_remote_local}
\alias{is_remote_local}
\title{Tests whether the remote is a local host or not.}
\usage{
is_remote_local(remote)
}
\arguments{
\item{remote}{A putative remote machine. This function will return true if \code{remote} is \code{NULL}, \code{NA}, or \code{""}.}
}
\description{
Tests whether the remote is a local host or not.
}
