% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{ls.remote}
\alias{ls.remote}
\title{View the contents of a directory remotely}
\usage{
ls.remote(path, remote = NULL, verbose = F)
}
\arguments{
\item{path}{Path of the directory.}

\item{remote}{Remote machine specification for ssh, in format such as \code{user@server} that does not 
require interactive password entry. For local execution, pass an empty string "" (default).}

\item{verbose}{If \code{TRUE} prints the command.}
}
\description{
View the contents of a directory remotely
}

