% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub.R
\name{manual.server.config}
\alias{manual.server.config}
\title{Use a one-time manual server config}
\usage{
manual.server.config(remote, src.dir, remote.dir)
}
\arguments{
\item{remote}{Remote machine specification for ssh, in format such as \code{user@server} that does not
require interactive password entry.}

\item{src.dir}{A directory on the local machine in which to store temporary files.}

\item{remote.dir}{A directory on the remote machine in which to store temporary files.}
}
\value{
A server configuration object
}
\description{
Use a one-time manual server config
}

