% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{mem.usage}
\alias{mem.usage}
\title{Measure the resident memory usage of a process.}
\usage{
mem.usage(pid = Sys.getpid())
}
\arguments{
\item{pid}{Process ID (default is the current process id).}
}
\value{
The resident memory usage in KB.
}
\description{
Returns the memory usage in KB of a process with the specified process id. By
default, returns the memory usage of the current R process. This can be used
to measure and log the memory usage of the R process during script execution.
}
\examples{
\dontrun{
mem.usage()
# [1] 37268
}
}

