% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub.R
\name{override.qsub.config}
\alias{override.qsub.config}
\title{Create a new qsub configuration object from an old qsub configuration.}
\usage{
override.qsub.config(qsub.config, ...)
}
\arguments{
\item{qsub.config}{The old qsub configuration to be modified.}

\item{...}{Any parameters to be overridden.}
}
\value{
A new qsub configuration object.
}
\description{
Create a new qsub configuration object from an old qsub configuration.
}

