% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub_config.R
\name{override_qsub_config}
\alias{override_qsub_config}
\title{Create a new qsub configuration object from an old qsub configuration.}
\usage{
override_qsub_config(qsub_config = NULL, ...)
}
\arguments{
\item{qsub_config}{The qsub_config to be overridden. If NULL, will attempt to retrieve a default qsub_config.}

\item{...}{Any parameters to be overridden.}
}
\value{
A new qsub_config object.
}
\description{
Create a new qsub configuration object from an old qsub configuration.
}
\examples{
\dontrun{
qsub_config <- create_qsub_config(
  remote = "myserver",
  local_tmp_path = "/home/myuser/workspace/.r2gridengine",
  remote_tmp_path = "/scratch/myuser/.r2gridengine"
)
qsub_lapply(1:10, function(x) x + 1, qsub_config = qsub_config)

qsub_config2 <- override_qsub_config(qsub_config, remote = "yourserver")
qsub_lapply(1:10, function(x) x + 1, qsub_config = qsub_config)
}
}
\seealso{
\code{\link{set_default_qsub_config}}
}

