% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qacct.R
\name{qacct_direct_remote}
\alias{qacct_direct_remote}
\title{Run qacct on remote}
\usage{
qacct_direct_remote(remote, job_id)
}
\arguments{
\item{remote}{The remote}

\item{job_id}{The job_id of the job}
}
\description{
Run qacct on remote
}
