% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qacct.R
\name{qacct_remote}
\alias{qacct_remote}
\title{Run qacct on remote}
\usage{
qacct_remote(job_id, remote = FALSE)
}
\arguments{
\item{job_id}{The job_id of the job}

\item{remote}{Remote machine specification for ssh, in format such as \code{user@server} that does not
require interactive password entry. For local execution, pass \code{FALSE} (default). For
execution on the default qsub config remote, use \code{TRUE}.}
}
\description{
Run qacct on remote
}
