% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstat_j.R
\name{qstat_j_remote}
\alias{qstat_j_remote}
\title{Run qstat on remote}
\usage{
qstat_j_remote(job_id, remote = FALSE)
}
\arguments{
\item{job_id}{The job_id of the job}

\item{remote}{Remote machine specification for ssh, in format such as \code{user@server} that does not
require interactive password entry. For local execution, pass \code{FALSE} (default). For
execution on the default qsub config remote, use TRUE.}
}
\description{
Run qstat on remote
}
