% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qacct.R, R/ssh.R
\name{qstat_remote}
\alias{qstat_remote}
\alias{qstat_remote}
\title{Run qstat on remote}
\usage{
qstat_remote(remote = NULL, verbose = F)

qstat_remote(remote = NULL, verbose = F)
}
\arguments{
\item{remote}{The remote}

\item{verbose}{If \code{TRUE} prints the command.}

\item{job_id}{The job_id of the job}

\item{remote}{Remote machine specification for ssh, in format such as \code{user@server} that does not
require interactive password entry. For local execution, pass an empty string "" (default).}
}
\description{
Run qstat on remote

Show the status of Grid Engine jobs and queues
}
