% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qsub.R
\name{qsub.configuration}
\alias{qsub.configuration}
\title{Create a qsub configuration object.}
\usage{
qsub.configuration(name = "R2PRISM", num.cores = 1, memory = "1G",
  verbose = F, remote = "prism", src.dir = "/tmp",
  remote.dir = "/scratch/irc/personal/robrechtc/tmp",
  tmp.foldername = paste0(name, "-", random::randomStrings(n = 1, len = 10)[1,
  ]), wait = T, remove.tmpdirs = T)
}
\arguments{
\item{name}{The name of the execution}

\item{num.cores}{The number of cores to allocate per task}

\item{memory}{The memory to allocate per core.}

\item{verbose}{Print out any ssh commands}

\item{remote}{Remote machine specification for ssh, in format such as \code{user@server} that does not
require interactive password entry.}

\item{src.dir}{A directory on the local machine in which to store temporary files.}

\item{remote.dir}{A directory on the remote machine in which to store temporary files.}

\item{tmp.foldername}{A temporary name for this execution.}

\item{wait}{If \code{TRUE}, will wait until the execution has finished by periodically checking the job status.}

\item{remove.tmpdirs}{If \code{TRUE}, will remove everything that was created related to this execution at the end.}
}
\description{
Create a qsub configuration object.
}

