% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub.R
\name{qsub.lapply}
\alias{qsub.lapply}
\title{Apply a Function over a List or Vector on PRISM!}
\usage{
qsub.lapply(X, FUN, qsub.config = qsub.configuration(),
  qsub.environment = NULL)
}
\arguments{
\item{X}{A vector (atomic or list) or an expression object. Other objects (including classed objects) will be coerced by base::as.list.}

\item{FUN}{The function to be applied to each element of X.}

\item{qsub.config}{The configuration to use for this execution.}

\item{qsub.environment}{\code{NULL}, a character vector or an environment.
Specifies what data and functions will be uploaded to the server.}
}
\description{
Apply a Function over a List or Vector on PRISM!
}
\examples{
\dontrun{
qsub.lapply(
  X=seq_len(3),
  FUN=function(i) { Sys.sleep(1); i+1 },
  qsub.config=qsub.configuration(
    remote.dir="/scratch/irc/personal/robrechtc/tmp")
)

promise <- qsub.lapply(
  X=seq_len(3),
  FUN=function(i) { Sys.sleep(1); i+1 },
  qsub.config=qsub.configuration(
    remote.dir="/scratch/irc/personal/robrechtc/tmp",
    wait=F)
)
qsub.retrieve(promise)
}
}

