% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{qsub}
\alias{qsub}
\title{Running Commands Remotely on Gridengine Clusters}
\description{
Run `lapply()` calls in parallel by submitting them to
'gridengine' clusters using the 'qsub' command.
}
