% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub.R
\name{qsub.retrieve}
\alias{qsub.retrieve}
\title{Retrieve the results of a qsub execution.}
\usage{
qsub.retrieve(qsub.config, wait = T)
}
\arguments{
\item{qsub.config}{The qsub configuration of class \code{qsub_configuration}, as returned by any qsub execution}

\item{wait}{If \code{TRUE}, wait until the execution has finished in order to return the results, else returns \code{NULL} if execution is not finished.}
}
\description{
Retrieve the results of a qsub execution.
}

