% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub_run.R
\name{qsub_run}
\alias{qsub_run}
\title{Calculate the results of a function on PRISM!}
\usage{
qsub_run(FUN, qsub_config = NULL, qsub_environment = NULL, ...)
}
\arguments{
\item{FUN}{the function to be executed.}

\item{qsub_config}{The configuration to use for this execution.}

\item{qsub_environment}{\code{NULL}, a character vector or an environment.
Specifies what data and functions will be uploaded to the server.}

\item{...}{optional arguments to FUN.}
}
\description{
Calculate the results of a function on PRISM!
}
\seealso{
\code{\link{create_qsub_config}}, \code{\link{set_default_qsub_config}}
}

