% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{rm_remote}
\alias{rm_remote}
\title{Remove a file or folder}
\usage{
rm_remote(path, remote, recursive = FALSE, force = FALSE, verbose = FALSE)
}
\arguments{
\item{path}{Path of the file/folder}

\item{remote}{Remote machine specification for ssh, in format such as \code{user@server} that does not
require interactive password entry. For local execution, pass \code{FALSE} (default). For
execution on the default qsub config remote, use \code{TRUE}.}

\item{recursive}{Whether to work recursively}

\item{force}{Whether to force removal}

\item{verbose}{If \code{TRUE} prints the command.}
}
\description{
Remove a file or folder
}
