% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{run_withwarn}
\alias{run_withwarn}
\title{Functions to run commands remotely via \code{ssh} and capture output.}
\usage{
run_withwarn(expr)
}
\arguments{
\item{expr}{Expression to be evaluated.}
}
\description{
\code{run_withwarn} - Evaluates the expression (e.g. a function call) and returns the
result with additional atributes:
\itemize{
\item num.warnings - number of warnings occured during the evaluation
\item last.message - the last warning message
}
Otherwise, \code{run_withwarn} is similar to \code{base::supressWarnings}
}

