% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub.R
\name{server.config.from.file}
\alias{server.config.from.file}
\title{Read a server config from file}
\usage{
server.config.from.file(file = "~/.local/share/R2PRISM/config")
}
\arguments{
\item{file}{The default location to store the R2PRISM config file.}
}
\value{
A server configuration object
}
\description{
Read a server config from file
}

