% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub_config.R
\name{test_qsub_config}
\alias{test_qsub_config}
\title{Tests whether the passed object is a qsub_config object.}
\usage{
test_qsub_config(object)
}
\arguments{
\item{object}{The object to be tested}
}
\description{
Tests whether the passed object is a qsub_config object.
}
