% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{write_remote}
\alias{write_remote}
\title{Write to a file remotely}
\usage{
write_remote(x, path, remote, verbose = FALSE)
}
\arguments{
\item{x}{The text to write to the file.}

\item{path}{Path of the file.}

\item{remote}{Remote machine specification for ssh, in format such as \code{user@server} that does not
require interactive password entry. For local execution, pass an empty string "" (default).}

\item{verbose}{If \code{TRUE} prints the command.}
}
\description{
Write to a file remotely
}
