% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{cat_remote}
\alias{cat_remote}
\title{Read from a file remotely}
\usage{
cat_remote(path, remote = NULL, verbose = FALSE)
}
\arguments{
\item{path}{Path of the file.}

\item{remote}{Remote machine specification for ssh, in format such as \code{user@server} that does not
require interactive password entry. For local execution, pass an empty string "" (default).}

\item{verbose}{If \code{TRUE} prints the command.}
}
\description{
Read from a file remotely
}
