% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub_config.R
\name{create_qsub_config}
\alias{create_qsub_config}
\title{Create a qsub configuration object.}
\usage{
create_qsub_config(
  # server settings
  remote,
  local_tmp_path,
  remote_tmp_path,

  # execution parameters
  name = "R2PRISM",
  num_cores = 1,
  memory = "4G",
  max_running_tasks = NULL,
  max_wall_time = "01:00:00",

  # pre-execution parameters
  r_module = "R",
  execute_before = NULL,
  verbose = FALSE,

  # post-execution parameters
  wait = TRUE,
  remove_tmp_folder = TRUE,
  stop_on_error = TRUE
)
}
\arguments{
\item{remote}{Remote machine specification for ssh, in format such as \code{user@server}
that does not require interactive password entry.}

\item{local_tmp_path}{A directory on the local machine in which to store temporary files. Should not contain a tilde ('~').}

\item{remote_tmp_path}{A directory on the remote machine in which to store temporary files. Should not contain a tilde ('~').}

\item{name}{The name of the execution.}

\item{num_cores}{The number of cores to allocate per task.}

\item{memory}{The memory to allocate per core.}

\item{max_running_tasks}{limit concurrent array job task execution.}

\item{max_wall_time}{The maximum time each task is allowed to run.}

\item{r_module}{The R module to use.}

\item{execute_before}{Commands to execute in the shell before running R.}

\item{verbose}{Print out any ssh commands.}

\item{wait}{If \code{TRUE}, will wait until the execution has finished by periodically checking the job status.}

\item{remove_tmp_folder}{If \code{TRUE}, will remove everything that was created related to this execution at the end.}

\item{stop_on_error}{If \code{TRUE}, will stop when an error occurs, else returns a NA for errored instances.}
}
\value{
A qsub configuration object.
}
\description{
Create a qsub configuration object.
}
\examples{
\dontrun{
qsub_config <- create_qsub_config(
  remote = "myserver",
  local_tmp_path = "/home/myuser/workspace/.r2gridengine",
  remote_tmp_path = "/scratch/myuser/.r2gridengine"
)
qsub_lapply(1:10, function(x) x + 1, qsub_config = qsub_config)

set_default_qsub_config(qsub_config, permanent = T)
qsub_lapply(1:10, function(x) x + 1)
}
}
\seealso{
\code{\link{qsub_lapply}}, \code{\link{set_default_qsub_config}}
}
