% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstat_j.R
\name{qstat_j_remote}
\alias{qstat_j_remote}
\title{Run qstat on remote}
\usage{
qstat_j_remote(remote, job_id)
}
\arguments{
\item{remote}{The remote}

\item{job_id}{The job_id of the job}
}
\description{
Run qstat on remote
}
