% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{qstat_remote}
\alias{qstat_remote}
\title{Show the status of Grid Engine jobs and queues}
\usage{
qstat_remote(remote = NULL, verbose = FALSE)
}
\arguments{
\item{remote}{Remote machine specification for ssh, in format such as \code{user@server} that does not
require interactive password entry. For local execution, pass an empty string "" (default).}

\item{verbose}{If \code{TRUE} prints the command.}
}
\description{
Show the status of Grid Engine jobs and queues
}
