% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub.R
\name{qsub.configuration}
\alias{qsub.configuration}
\title{Create a qsub configuration object.}
\usage{
qsub.configuration(r.module = "R", name = "R2PRISM", num.cores = 1,
  memory = "4G", verbose = F, tmp.foldername = paste0(name, "-",
  random::randomStrings(n = 1, len = 10)[1, ]), wait = T,
  remove.tmpdirs = T, stop.on.error = T, max.tasks = NULL,
  server.config = server.config.from.file())
}
\arguments{
\item{r.module}{The R module to use}

\item{name}{The name of the execution}

\item{num.cores}{The number of cores to allocate per task}

\item{memory}{The memory to allocate per core.}

\item{verbose}{Print out any ssh commands}

\item{tmp.foldername}{A temporary name for this execution.}

\item{wait}{If \code{TRUE}, will wait until the execution has finished by periodically checking the job status.}

\item{remove.tmpdirs}{If \code{TRUE}, will remove everything that was created related to this execution at the end.}

\item{stop.on.error}{If \code{TRUE}, will stop when an error occurs, else returns a NA for errored instances.}

\item{max.tasks}{The maximum number of tasks to spawn}

\item{server.config}{A server configuration file}
}
\description{
Create a qsub configuration object.
}

