% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qsub.R
\name{qsublapply}
\alias{qsublapply}
\title{Apply a Function over a List or Vector on PRISM!}
\usage{
qsublapply(X, FUN, qsub.remote = "prism", qsub.memory = "1G",
  qsub.cores = 1, qsub.name = "R2PRISM", qsub.src.dir = "/tmp",
  qsub.remote.dir = "/scratch/irc/personal/robrechtc/tmp",
  qsub.tmp.foldername = paste0("R2PRISM-", random::randomStrings(n = 1, len =
  10)[1, ]), qsub.remove.tmpdirs = T, qsub.verbose = F)
}
\arguments{
\item{X}{a vector (atomic or list) or an expression object. Other objects (including classed objects) will be coerced by base::as.list.}

\item{FUN}{the function to be applied to each element of X.}

\item{qsub.remote}{ssh alias for PRISM}

\item{qsub.memory}{memory to allocate for each task.}

\item{qsub.cores}{number of cores per task.}

\item{qsub.name}{name given to the task.}

\item{qsub.src.dir}{a local temporary folder.}

\item{qsub.remote.dir}{a remote temporary folder.}

\item{qsub.tmp.foldername}{a random tmp folder name.}

\item{qsub.remove.tmpdirs}{remove the temporary folders at the end of this procedure?}

\item{qsub.verbose}{When \code{TRUE}, prints out the commands which are run.}
}
\description{
Apply a Function over a List or Vector on PRISM!
}
\examples{
\dontrun{
# X=seq_len(100)
# FUN=function(i) { Sys.sleep(1); i }
# qsublapply(X, FUN, qsub.remote.dir = "/scratch/irc/personal/robrechtc/tmp")
}
}

